* commit 'cbec51f5f2b9c425ce551405518e48d4046adb35':
  SHA1_MessageDigestImpl should work for inputs of 256 Megabytes

+++ luni/src/main/java/org/apache/harmony/security/provider/crypto/SHA1_MessageDigestImpl.java
@@ -42,7 +42,7 @@ public class SHA1_MessageDigestImpl extends MessageDigestSpi implements Cloneabl
     private byte[] oneByte;      // one byte buffer needed to use in engineUpdate(byte)
                                  // having buffer as private field is just optimization
 
-    private int messageLength;   // total length of bytes supplied by user
+    private long messageLength;   // total length of bytes supplied by user
 
 
     /**
